+++ timeline/templatetags/timeline_tags.py
@@ -36,9 +36,9 @@ def get_timestamp(event):
 
 
 @register.simple_tag
-def get_event_description(event, request=None):
+def get_event_description(event, plugin_lookup, request=None):
     """Return printable version of event description"""
-    return timeline.get_event_description(event, request)
+    return timeline.get_event_description(event, plugin_lookup, request)
 
 
 @register.simple_tag
@@ -57,7 +57,7 @@ def get_plugin_lookup():
 
 
 @register.simple_tag
-def get_app_icon_html(plugin_lookup, event):
+def get_app_icon_html(event, plugin_lookup):
     """Return icon link HTML for app by plugin lookup"""
     url = None
     url_kwargs = {}